Refactor Data Reduction Manager into It's Own Class
